pyproject.toml
[project]
name = "chatbot-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "groq>=0.31.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "openai>=1.107.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "tqdm>=4.67.1",
]
